Refactor core modules into temporal_gradient package
