+++ temporal_gradient/__init__.py
@@ -1,0 +1,1 @@
+"""Temporal Gradient package."""
