+++ temporal_gradient/clock/__init__.py
@@ -1,0 +1,3 @@
+from .chronos import ClockRateModulator
+
+__all__ = ["ClockRateModulator"]
