+++ temporal_gradient/contracts/memory.py
@@ -1,0 +1,16 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class MemoryEncodingDecision:
+    should_encode: bool
+    psi: float
+    threshold: float
+    initial_strength: float
+
+
+@dataclass(frozen=True)
+class MemoryDecaySnapshot:
+    memory_id: str
+    current_strength: float
+    pruned: bool
