+++ temporal_gradient/memory/__init__.py
@@ -1,0 +1,3 @@
+from .decay import DecayEngine, EntropicMemory, S_MAX, initial_strength_from_psi, should_encode
+
+__all__ = ["DecayEngine", "EntropicMemory", "S_MAX", "initial_strength_from_psi", "should_encode"]
