+++ temporal_gradient/salience/__init__.py
@@ -1,0 +1,21 @@
+from .pipeline import (
+    CodexNoveltyAdapter,
+    CodexValueAdapter,
+    KeywordImperativeValue,
+    NoveltyScorer,
+    RollingJaccardNovelty,
+    SalienceComponents,
+    SaliencePipeline,
+    ValueScorer,
+)
+
+__all__ = [
+    "CodexNoveltyAdapter",
+    "CodexValueAdapter",
+    "KeywordImperativeValue",
+    "NoveltyScorer",
+    "RollingJaccardNovelty",
+    "SalienceComponents",
+    "SaliencePipeline",
+    "ValueScorer",
+]
